# api/openapi/spec.json
{
...
  "info": {
    "description": "Secure AI agent gateway API",
    "title": "Sigil Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/chat/stream": {
      "post": {
        "description": "Send a message and receive a streaming response. Set Accept: text/event-stream for SSE, otherwise receives a JSON object containing an events array.",
        "operationId": "chat-stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "content": {
                    "description": "Message content",
                    "minLength": 1,
                    "type": "string"
                  },
                  "session_id": {
                    "description": "Optional session to resume",
                    "type": "string"
                  },
                  "workspace_id": {
                    "description": "Target workspace",
                    "type": "string"
                  }
                },
                "required": [
                  "content",
                  "workspace_id"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "events": {
                      "description": "Collected events with type and data",
                      "items": {
                        "properties": {
                          "data": {
                            "description": "Event payload"
                          },
                          "event": {
                            "description": "Event type (text_delta, session_id, error, done)",
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              },
              "text/event-stream": {
                "schema": {
                  "description": "Server-sent event stream",
                  "type": "string"
                }
              }
            },
            "description": "Streaming response (SSE or JSON depending on Accept header)"
          },
          "422": {
            "description": "Validation error (missing content)"
          },
          "503": {
            "description": "Stream handler not configured"
          }
        },
        "summary": "Stream a chat response via SSE",
        "tags": [
          "chat"
        ]
      }
    },
...
  }
}